controller/pyproject.toml
[tool.poetry]
name = "plant-diagnosis-controller"
version = "0.1.0"
description = ""
authors = ["xkurozaru <kurozaru.210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.0.1+cpu", source = "pytorch"}
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
torchvision = "^0.15.2"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.21"
cuid2 = "^2.0.0"
python-jose = "^3.3.0"
mysql-connector-python = "^8.1.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
